refactor and add more complete comments
