+++ package.json
@@ -54,6 +54,7 @@
     "selenium-webdriver": "^4.0.0-rc-1",
     "tailwindcss": "^2.2.16",
     "ts-jest": "^27.0.7",
+    "typescript": "^4.5.2",
     "web-ext": "^6.4.0",
     "webextension-polyfill": "^0.8.0"
   },
