+++ src/pixelHuntStudy.ts
@@ -6,6 +6,7 @@ import browser from "webextension-polyfill";
 
 import * as pixelHuntPings from "../src/generated/pings";
 import * as facebookPixel from "../src/generated/facebookPixel";
+import * as userJourney from "../src/generated/userJourney.js";
 
 const fbHostname = ["www.facebook.com"];
 // @ts-ignore
@@ -14,9 +15,25 @@ const enableDevMode = Boolean(__ENABLE_DEVELOPER_MODE__);
 if (enableDevMode) {
   fbHostname.push("localhost");
 }
-// responds to browser.webRequest.onCompleted events
-export async function fbPixelListener(details) {
-  // Facebook pixels live at `*://www.facebook.com/tr/`
+/**
+  * Responds to browser.webRequest.onCompleted events.
+  *
+  * @param {browser.WebRequest.OnBeforeRequestDetailsTypes} - details for the web request.
+  *
+  */
+export function fbPixelListener(details: browser.WebRequest.OnBeforeRequestDetailsType) {
+  handlePixel(details).catch(err => console.error("Facbook Pixel Hunt Listener Error:", err));
+}
+
+/**
+ * Internal async handler for Facebook pixels.
+ *
+ * Note: Chrome and Firefox both require either `void` or `Promise<BlockingResponse>` here, because this
+ * API *might* block content. We just want to handle this async, since we depend on async WebExtension APIs.
+ *
+ * @param {browser.WebRequest.OnBeforeRequestDetailsTypes} - details for the web request.
+ */
+async function handlePixel(details: browser.WebRequest.OnBeforeRequestDetailsType) {
   const url = new URL(details.url);
   let originUrl = undefined;
   if (details.originUrl) {
@@ -24,6 +41,7 @@ export async function fbPixelListener(details) {
   }
   const tabId = details.tabId;
 
+  // Facebook pixels live at `*://www.facebook.com/tr/`
   if (fbHostname.includes(url.hostname) && url.pathname.match(/^\/tr/)) {
     facebookPixel.url.setUrl(url);
 
@@ -48,7 +66,7 @@ export async function fbPixelListener(details) {
 
     // Attempt to associate this pixel tracker sighting with a WebScience Page ID.
     const pageVisits = (await browser.storage.local.get("pageVisits"))["pageVisits"];
-    let pageId;
+    let pageId: string;
     for (const visit of pageVisits) {
       const visitUrl = new URL(visit.url);
       if (visitUrl.origin === originUrl.origin && visit.tabId === tabId) {
@@ -86,4 +104,72 @@ export async function fbPixelListener(details) {
       pixelHuntPings.fbpixelhuntPixel.submit();
     }
   }
+}
+
+/**
+ * Listen for page navigation ("user journey") events.
+ *
+ * @param {Object} pageData - WebScience page data details.
+ */
+export async function pageDataListener(pageData) {
+  if (enableDevMode) {
+    //
+    // FIXME it would be preferable to get this straight from Glean, but unfortunately it does not seem to be
+    // holding more than one ping at a time in its local storage when submission is disabled.
+    // TODO file issue to follow up.
+
+    const pageNavigationPings = (await browser.storage.local.get("pageNavigationPings"))["pageNavigationPings"];
+    // If this storage object already exists, append to it.
+    const result = pageData;
+    if (Array.isArray(pageNavigationPings)) {
+      pageNavigationPings.push(result);
+
+      await browser.storage.local.set({ pageNavigationPings });
+    } else {
+      await browser.storage.local.set({ "pageNavigationPings": [result] });
+    }
+  } else {
+    if (!pageData.pageId) {
+      console.warn("No pageID assigned by pageNavigation:", pageData);
+    }
+    userJourney.pageId.set(pageData.pageId);
+    userJourney.attentionDuration.set(pageData.attentionDuration);
+    userJourney.audioDuration.set(pageData.audioDuration);
+    userJourney.maxRelativeScrollDepth.set(pageData.maxRelativeScrollDepth);
+    userJourney.pageVisitStartTime.set(pageData.pageVisitStartTime);
+    userJourney.pageVisitStopTime.set(pageData.pageVisitStopTime);
+    userJourney.referrer.setUrl(pageData.referrer);
+    userJourney.url.setUrl(pageData.url);
+
+    pixelHuntPings.fbpixelhuntJourney.submit();
+  }
+}
+
+/**
+ * Listen for page visit start and record it, so we can match up facebook pixel events with pages in the user journey.
+ * This adds a new entry to local storage regarding this page visit.
+ *
+ * @param {Object} pageVisit - WebScience page visit details.
+ */
+export async function pageVisitStartListener(pageVisit) {
+  const pageVisits = (await browser.storage.local.get("pageVisits"))["pageVisits"];
+  // If this storage object already exists, append to it.
+  if (Array.isArray(pageVisits)) {
+    pageVisits.push(pageVisit);
+
+    await browser.storage.local.set({ pageVisits });
+  } else {
+    await browser.storage.local.set({ "pageVisits": [pageVisit] });
+  }
+}
+
+/**
+ * Listen for page visit stop. This removes any entries from local storage regarding this page visit.
+ *
+ * @param {Object} pageVisit - WebScience page visit details.
+ */
+export async function pageVisitStopListener(pageVisit) {
+  let pageVisits = (await browser.storage.local.get("pageVisits"))["pageVisits"];
+  pageVisits = pageVisits.filter(a => a.pageId !== pageVisit.pageId)
+  await browser.storage.local.set({ pageVisits });
 }
